+++ lib/rubber.dart
@@ -6,5 +6,7 @@ library rubber;
 export 'src/animation_controller.dart';
 export 'src/bottom_sheet.dart';
 export 'src/spring_simulation.dart';
+export 'src/spring_description/damping_ratio.dart';
+export 'src/spring_description/stiffness.dart';
 
 export 'package:flutter/physics.dart';