+++ lib/src/spring_description/damping_ratio.dart
@@ -1,0 +1,6 @@
+class DampingRatio {
+  static const HIGH_BOUNCY = 0.2;
+  static const MEDIUM_BOUNCY = 0.5;
+  static const LOW_BOUNCY = 0.75;
+  static const NO_BOUNCY = 1.0;
+}