+++ lib/src/spring_description/stiffness.dart
@@ -1,0 +1,6 @@
+class Stiffness {
+  static const HIGH = 10000.0;
+  static const MEDIUM = 1500.0;
+  static const LOW = 200.0;
+  static const VERY_LOW = 50.0;
+}